% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello World}
\usage{
hello()
}
\value{
"Hello World"
}
\description{
This **wonderful** function will print "Hello World"
}
\details{
The user does not need to provide any argument

```{r examples}
hello()
```
}
\examples{

hello()


}
